# backend-api.openapi.yaml
openapi: 3.0.3
info:
  title: Car Service Admin API
  version: 2.0.0
  description: |
    Backend contract for the Car Service Admin dashboard.

    The frontend currently covers authentication, onboarding, dashboard widgets,
    bookings, payments, payment exports, reports, storage inventory, business
    settings, billing account verification, notification preferences, inbox,
    feedback, help center content, and location search.

    API JSON uses camelCase. The recommended database schema uses snake_case.
    Dates should be stored and exchanged as ISO-8601 strings. The frontend can
    format those dates for display.
  contact:
    name: Car Service Admin
servers:
  - url: http://localhost:3333
    description: Local backend development
  - url: https://api.car-service-admin.example.com
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /auth/sign-in:
    post:
      tags: [Auth]
      operationId: signIn
      summary: Sign in as a client, business user, or admin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInDto"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/sign-up:
    post:
      tags: [Auth]
      operationId: signUpBusiness
      summary: Register a new business workspace and owner account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessSignUpDto"
      responses:
        "201":
          description: Business account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Rotate a refresh token and return a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenDto"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get the current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/sign-out:
    post:
      tags: [Auth]
      operationId: signOut
      summary: Revoke the current refresh token/session
      responses:
        "204":
          description: Signed out
        "401":
          $ref: "#/components/responses/Unauthorized"
  /account/profile:
    get:
      tags: [Account]
      operationId: getAccountProfile
      summary: Get personal profile and address settings
      responses:
        "200":
          description: Account profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountProfile"
    patch:
      tags: [Account]
      operationId: updateAccountProfile
      summary: Update personal profile and address settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountProfileUpdateDto"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountProfile"
        "400":
          $ref: "#/components/responses/ValidationError"
  /account/security:
    get:
      tags: [Account]
      operationId: getSecuritySettings
      summary: Get security preferences and MFA state
      responses:
        "200":
          description: Security settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecuritySettings"
    patch:
      tags: [Account]
      operationId: updateSecuritySettings
      summary: Update security preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecuritySettingsUpdateDto"
      responses:
        "200":
          description: Updated security settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecuritySettings"
  /account/mfa:
    put:
      tags: [Account]
      operationId: configureMfa
      summary: Enable, update, or disable multi-factor authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaUpdateDto"
      responses:
        "200":
          description: Updated MFA settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MfaSettings"
        "400":
          $ref: "#/components/responses/ValidationError"
  /account/notification-preferences:
    get:
      tags: [Notifications]
      operationId: getNotificationPreferences
      summary: Get account toast/action notification preferences
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
    put:
      tags: [Notifications]
      operationId: updateNotificationPreferences
      summary: Replace account notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferencesUpdateDto"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
...
  /bookings:
    get:
      tags: [Bookings]
      operationId: listBookings
      summary: List bookings with filters, search, sort, and pagination
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/BookingStatus"
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: search
          description: Matches booking id, service name, customer name, car make, or payment id.
          schema:
            type: string
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [serviceName, name, carMake, startDate, endDate, status, price]
        - $ref: "#/components/parameters/SortOrderQuery"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Paginated bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedBookings"
    post:
      tags: [Bookings]
      operationId: createBooking
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingCreateDto"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          $ref: "#/components/responses/ValidationError"
  /bookings/{bookingId}:
    get:
      tags: [Bookings]
      operationId: getBooking
      summary: Get booking details
      parameters:
        - $ref: "#/components/parameters/BookingIdPath"
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Bookings]
      operationId: updateBooking
      summary: Update booking fields
      parameters:
        - $ref: "#/components/parameters/BookingIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdateDto"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Bookings]
      operationId: deleteBooking
      summary: Delete a booking
      parameters:
        - $ref: "#/components/parameters/BookingIdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /bookings/{bookingId}/status:
    patch:
      tags: [Bookings]
      operationId: updateBookingStatus
      summary: Accept, decline, complete, or move one booking
      parameters:
        - $ref: "#/components/parameters/BookingIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdateDto"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /bookings/batch/status:
    patch:
      tags: [Bookings]
      operationId: updateBookingStatuses
      summary: Bulk update booking statuses for selected rows
      x-queue:
        name: notification-dispatch
        reason: Send accepted/declined notifications without blocking the booking update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingBatchStatusUpdateDto"
      responses:
        "200":
          description: Batch status update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchMutationResult"
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /billing/accounts:
    get:
      tags: [Billing]
      operationId: listBillingAccounts
      summary: List business billing/bank accounts
      responses:
        "200":
          description: Billing accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillingAccount"
    post:
      tags: [Billing]
      operationId: createBillingAccount
      summary: Create a billing account draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountCreateDto"
      responses:
        "201":
          description: Billing account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingAccount"
        "400":
          $ref: "#/components/responses/ValidationError"
  /billing/accounts/{billingAccountId}:
    get:
      tags: [Billing]
      operationId: getBillingAccount
      summary: Get billing account details
      parameters:
        - $ref: "#/components/parameters/BillingAccountIdPath"
      responses:
        "200":
          description: Billing account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingAccount"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Billing]
      operationId: updateBillingAccount
      summary: Update billing account draft details
      parameters:
        - $ref: "#/components/parameters/BillingAccountIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountUpdateDto"
      responses:
        "200":
          description: Updated billing account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingAccount"
        "400":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Billing]
      operationId: deleteBillingAccount
      summary: Delete a billing account draft
      parameters:
        - $ref: "#/components/parameters/BillingAccountIdPath"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /billing/accounts/{billingAccountId}/submit:
    post:
      tags: [Billing]
      operationId: submitBillingAccount
      summary: Submit billing account for verification
      x-queue:
        name: billing-verification
        reason: Bank account verification can be slow and should be retried asynchronously.
      parameters:
        - $ref: "#/components/parameters/BillingAccountIdPath"
      responses:
        "202":
          description: Billing account submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingAccount"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageQuery:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitQuery:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortOrderQuery:
      in: query
      name: sortOrder
      schema:
        type: string
        enum: [asc, desc]
        default: asc
    BookingIdPath:
      in: path
      name: bookingId
      required: true
      schema:
        type: string
...
    BillingAccountIdPath:
      in: path
      name: billingAccountId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication is missing, invalid, or expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: Request conflicts with current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      required: [statusCode, code, message]
      properties:
        statusCode:
          type: integer
          example: 400
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Please enter a valid email address.
        details:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorDetail"
        requestId:
          type: string
          nullable: true
...
    BatchMutationResult:
      type: object
      required: [requested, succeeded, failed]
      properties:
        requested:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        failedIds:
          type: array
          items:
            type: string
    Role:
      type: string
      enum: [client, business, admin]
    SignInDto:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
        role:
          type: string
          enum: [client, business, admin]
        rememberMe:
          type: boolean
          default: false
    RefreshTokenDto:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          default: Bearer
        expiresIn:
          type: integer
          description: Access token lifetime in seconds.
          example: 900
        user:
          $ref: "#/components/schemas/User"
    BusinessSignUpDto:
      type: object
      required:
        - email
        - password
        - name
        - phoneNumber
        - country
        - city
        - commercialRegistrationNumber
        - taxIdentificationNumber
        - businessOwner
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[^A-Za-z0-9]).{12,}$"
        name:
          type: string
          minLength: 1
        phoneNumber:
          type: string
        country:
          type: string
        city:
          type: string
        commercialRegistrationNumber:
          type: string
          description: Correct spelling for the backend/database field.
        commercialRegistraionNumber:
          type: string
          deprecated: true
          description: Legacy frontend typo. Accept only as a transitional alias.
        taxIdentificationNumber:
          type: string
        businessOwner:
          $ref: "#/components/schemas/BusinessOwnerCreateDto"
...
    User:
      type: object
      required: [id, email, role, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        businessId:
          type: string
          format: uuid
          nullable: true
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        onboardingCompleted:
          type: boolean
        twoFactorEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AccountProfile:
      type: object
      required: [id, firstName, lastName, address]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        address:
          $ref: "#/components/schemas/Address"
        setupStatus:
          $ref: "#/components/schemas/SetupStatus"
    AccountProfileUpdateDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
          nullable: true
        phoneNumber:
          type: string
        address:
          $ref: "#/components/schemas/Address"
...
    SecuritySettings:
      type: object
      required: [preferences, mfa]
      properties:
        preferences:
          type: array
          items:
            $ref: "#/components/schemas/SecurityPreference"
        mfa:
          $ref: "#/components/schemas/MfaSettings"
    SecuritySettingsUpdateDto:
      type: object
      required: [preferences]
      properties:
        preferences:
          type: object
          additionalProperties:
            type: boolean
          example:
            loginAlerts: true
            newDeviceApproval: true
            sessionTimeout: true
            paymentExportConfirmation: false
...
    MfaSettings:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        method:
          $ref: "#/components/schemas/MfaMethod"
        phone:
          type: string
          nullable: true
        verifiedAt:
          type: string
          format: date-time
          nullable: true
    MfaUpdateDto:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        method:
          $ref: "#/components/schemas/MfaMethod"
        phone:
          type: string
          nullable: true
          pattern: "^\\+?[0-9\\s()-]{8,20}$"
...
    NotificationPreferences:
      type: object
      required: [preferences]
      properties:
        preferences:
          type: array
          items:
            $ref: "#/components/schemas/NotificationPreference"
    NotificationPreferencesUpdateDto:
      type: object
      required: [preferences]
      properties:
        preferences:
          type: object
          additionalProperties:
            type: boolean
          example:
            bookingAccepted: true
            bookingDeclined: true
            settingsSaved: true
            securitySettingsUpdated: true
            mfaUpdated: true
            exportGenerated: true
            storageItemSold: true
...
    BookingStatus:
      type: string
      enum: [Pending, Accepted, In-Progress, Completed, Cancelled]
...
    Booking:
      type: object
      required:
        - id
        - serviceName
        - name
        - carMake
        - carModel
        - carYear
        - startDate
        - endDate
        - status
        - price
      properties:
        id:
          type: string
          example: BK-10023
        businessId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        serviceName:
          type: string
        serviceIcon:
          type: string
          nullable: true
        ownerIcon:
          type: string
          nullable: true
        name:
          type: string
          description: Customer display name, kept to match the current frontend entity.
        carMake:
          type: string
        carModel:
          type: string
        carYear:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/BookingStatus"
        price:
          type: number
          format: double
        currency:
          type: string
          default: USD
        notes:
          type: string
          nullable: true
        serviceDetails:
          type: array
          items:
            $ref: "#/components/schemas/ServiceDetail"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookingCreateDto:
      type: object
      required: [serviceId, customerId, vehicleId, startDate, endDate]
      properties:
        serviceId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/BookingStatus"
        price:
          type: number
          format: double
        currency:
          type: string
          default: USD
        notes:
          type: string
        serviceDetails:
          type: array
          items:
            $ref: "#/components/schemas/ServiceDetail"
    BookingUpdateDto:
      type: object
      properties:
        serviceId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/BookingStatus"
        price:
          type: number
          format: double
        notes:
          type: string
        serviceDetails:
          type: array
          items:
            $ref: "#/components/schemas/ServiceDetail"
    BookingStatusUpdateDto:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/BookingStatus"
        reason:
          type: string
          nullable: true
    BookingBatchStatusUpdateDto:
      type: object
      required: [ids, status]
      properties:
        ids:
          type: array
          minItems: 1
          items:
            type: string
        status:
          $ref: "#/components/schemas/BookingStatus"
        reason:
          type: string
          nullable: true
    PaginatedBookings:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    BillingAccount:
      type: object
      required: [id, businessId, bankName, accountHolder, accountNumber, iban, swift, branch, billingEmail, status]
      properties:
        id:
          type: string
          format: uuid
        businessId:
          type: string
          format: uuid
        bankName:
          type: string
        accountHolder:
          type: string
        accountNumber:
          type: string
        iban:
          type: string
        swift:
          type: string
        branch:
          type: string
        billingEmail:
          type: string
          format: email
        status:
          $ref: "#/components/schemas/BillingAccountStatus"
        verificationNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        submittedAt:
          type: string
          format: date-time
          nullable: true
    BankAccountCreateDto:
      type: object
      required: [bankName, accountHolder, accountNumber, iban, swift, branch, billingEmail]
      properties:
        bankName:
          type: string
        accountHolder:
          type: string
          pattern: "^[A-Za-z\\s]+$"
        accountNumber:
          type: string
          minLength: 6
          maxLength: 34
          pattern: "^[0-9]+$"
        iban:
          type: string
          pattern: "^[A-Z]{2}\\d{2}[A-Z0-9]{11,30}$"
        swift:
          type: string
          pattern: "^[A-Z]{4}[A-Z]{2}[A-Z0-9]{2}([A-Z0-9]{3})?$"
        branch:
          type: string
          pattern: "^[A-Za-z0-9\\s.'&-]+$"
        billingEmail:
          type: string
          format: email
    BankAccountUpdateDto:
      type: object
      properties:
        bankName:
          type: string
        accountHolder:
          type: string
          pattern: "^[A-Za-z\\s]+$"
        accountNumber:
          type: string
          minLength: 6
          maxLength: 34
          pattern: "^[0-9]+$"
        iban:
          type: string
          pattern: "^[A-Z]{2}\\d{2}[A-Z0-9]{11,30}$"
        swift:
          type: string
          pattern: "^[A-Z]{4}[A-Z]{2}[A-Z0-9]{2}([A-Z0-9]{3})?$"
        branch:
          type: string
          pattern: "^[A-Za-z0-9\\s.'&-]+$"
        billingEmail:
          type: string
          format: email
...